requirements.txt
openai
langchain
python-dotenv
tiktoken